+++ tests/conftest.py
@@ -31,24 +31,25 @@ MEMORY_TESTCONFIGS = SubCache(
 # MEMORY_TESTCONFIGS.clear()
 
 
-# %% Datasets
-
-### Source of files is seaborn, markurutils just adds cut column
-
-# ==
-# == Load example data =================================================================
+# %% 
+# == Load Datasets =====================================================
+#' Source of files is seaborn, markurutils just adds cut column
 
 ### Cache function
 load_dataset = MEMORY_TESTCONFIGS.subcache(plst.load_dataset)
 
+### Load datasets
 DF_tips, dims_tips = load_dataset("tips", verbose=False)
 DF_fmri, dims_fmri = load_dataset("fmri", verbose=False)
 DF_qpcr, dims_qpcr = load_dataset("qpcr", verbose=False)
 
 
-# %% Arguments
+# %% 
+# == Arguments for plst.DataAnalysis ===================================
+#' Facet data such that it leads to presence of absence of empty groups
 
-### Args that make empty groups
+
+### Empty groups
 dims_withempty_tips = [
     dict(y="tip", x="day", hue="sex", col="smoker", row="time"),
     dict(y="tip", x="sex", hue="day", col="smoker", row="time"),
@@ -62,8 +63,7 @@ dims_withempty_tips = [
 ]
 
 
-### Args making sure they don't make empty groups
-# !! Don't add more, other tests assume 4 entries
+### Don't make empty groups
 dims_noempty_tips = [
     dict(y="tip", x="size-cut", hue="smoker", col="sex", row="time"),
     dict(y="tip", x="size-cut", hue="smoker", col="sex"),
@@ -87,12 +87,13 @@ dims_noempty_qpcr = [
     dict(y="FC", x="gene"),
 ]
 
-# %%  Combine for pytest.parametrize
+# %%  
+# == Make tuples (DF, dims) ============================================
+#' for pytest.parametrize
 
 zipped_noempty_tips = [(DF_tips, dim) for dim in dims_noempty_tips]
 zipped_noempty_fmri = [(DF_fmri, dim) for dim in dims_noempty_fmri]
 zipped_noempty_qpcr = [(DF_qpcr, dim) for dim in dims_noempty_qpcr]
-
 
 ### Paired Data (with subject)
 zipped_noempty_PAIRED = zipped_noempty_fmri + zipped_noempty_qpcr
@@ -104,10 +105,11 @@ zipped_noempty_ALL = (
 # len(zipped_noempty_ALL) # * -> 14 total tests
 
 
-# %% Dataanalysis objects
+# %% 
+# == Make Dataanalysis objects =========================================
 
 
-def get_DA_statistics(dataset: str = "qpcr") -> plst.DataAnalysis:
+def make_DA_statistics(dataset: str = "qpcr") -> plst.DataAnalysis:
     """Makes a DA object with every possible data stored in it
 
     :param dataset: "tips", "fmri", or "qpcr"
@@ -144,7 +146,7 @@ def get_DA_statistics(dataset: str = "qpcr") -> plst.DataAnalysis:
     return DA
 
 
-def get_DA_plot(dataset: str = "qpcr") -> plst.DataAnalysis:
+def make_DA_plot(dataset: str = "qpcr") -> plst.DataAnalysis:
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
         ### Load example data
@@ -158,17 +160,18 @@ def get_DA_plot(dataset: str = "qpcr") -> plst.DataAnalysis:
         return DA
 
 
-def get_DA_all(dataset: str) -> plst.DataAnalysis:
[…]
