+++ tests/joblib/testconfigs/get_DA_all/func_code.py
@@ -3,7 +3,7 @@ def get_DA_all(dataset: str) -> plst.DataAnalysis:
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
 
-        DA = get_DA_statistics(dataset)
+        DA = make_DA_statistics(dataset)
         DA.plot_box_swarm()
         plt.close()
         return DA
