+++ testing/make_htmlcov.sh
@@ -1,0 +1,4 @@
+
+
+### From within project root
+pytest tests -n 3 --cov --cov-report html:./testing/htmlcov --cov-config=./testing/.coveragerc 
