+++ .github/workflows/test_coverage.yml
@@ -17,7 +17,7 @@ jobs:
     - name: Install plotastic (in editable mode! Otherwise 0% coverage)
       run: pip install -e .[dev]
     - name: Run tests and collect coverage
-      run: pytest -n 4 tests --cov --cov-report=xml --cov-config=./testing/.coveragerc 
+      run: pytest tests --cov --cov-report=xml --cov-config=./testing/.coveragerc 
     - name: Upload coverage reports to Codecov with GitHub Action
       uses: codecov/codecov-action@v3
       with:
