+++ tests/multiplot_test.py
@@ -1,0 +1,7 @@
+#
+# %%
+
+import pytest
+
+# %%
+

[…]
