### 2023-11-26 24bdaba
tests for multiplots
