+++ tests/plotedits_test.py
@@ -70,8 +70,7 @@ zipped_tips = dac.add_zip_column(dac.zipped_noempty_tips, titles_zip)
 def test_edit_titles_replace(DF, dims, titles: dict):
     DA = plst.DataAnalysis(data=DF, dims=dims, verbose=False)
     if DA.factors_rowcol:  #' Need facetting, otherwise no axes
-        DA.plot()
-        DA.edit_titles_replace(titles=titles)
+        (DA.plot().edit_titles_replace(titles=titles))
     if __name__ == "__main__":
         plt.show()
     else:
@@ -81,5 +80,25 @@ def test_edit_titles_replace(DF, dims, titles: dict):
 if __name__ == "__main__":
     ipytest.run()
 
-#%%
+# %%
 
+
+@pytest.mark.parametrize("DF, dims", dac.zipped_ALL)
+def test_edit_xy_axis_labels(DF, dims):
+    DA = plst.DataAnalysis(data=DF, dims=dims, verbose=False)
+    DA.plot().edit_xy_axis_labels(
+        x="ui!",
+        x_lowest_row="rambazamba",
+        x_notlowest_row="FLOH",
+        y="Johannes",
+        y_leftmost_col="Gertrude",
+        y_notleftmost_col="Hä?",
+    )
+    if __name__ == "__main__":
+        plt.show()
+    else:
+        plt.close("all")
+
+
+if __name__ == "__main__":
+    ipytest.run()
