+++ tests/plotedits_test.py
@@ -1,3 +1,4 @@
+#
 # %%
 
 import pytest
@@ -10,13 +11,13 @@ import _DA_configs as dac
 
 
 # %%
-titles_zip = [
+titles_tips = [
     {("Lunch", "Male"): "blaa"},
     {("Male"): "blAA"},
     None,
     None,
 ]
-zipped_tips = dac.add_zip_column(dac.zipped_noempty_tips, titles_zip)
+zipped_tips = dac.add_zip_column(dac.zipped_noempty_tips, titles_tips)
 
 
 @pytest.mark.parametrize("DF, dims, axtitles", zipped_tips)
@@ -56,13 +57,13 @@ if __name__ == "__main__":
     ipytest.run()
 
 # %%
-titles_zip = [
+titles_tips = [
     ["sdfsfd", None, "dd", None],
     [None, "aa"],
     None,
     None,
 ]
-zipped_tips = dac.add_zip_column(dac.zipped_noempty_tips, titles_zip)
+zipped_tips = dac.add_zip_column(dac.zipped_noempty_tips, titles_tips)
 
 
 @pytest.mark.parametrize("DF, dims, titles", zipped_tips)
@@ -123,8 +124,145 @@ def test_edit_y_ticklabel_percentage(DF, dims):
     DA = plst.DataAnalysis(data=DF, dims=dims, verbose=False)
     DA.plot().edit_y_ticklabel_percentage(
         decimals_major=1,
-        decimals_minor=1, # !! Not working
+        decimals_minor=1,  # !! Not working
     )
+    if __name__ == "__main__":
+        plt.show()
+    else:
+        plt.close("all")
+
+
+if __name__ == "__main__":
+    ipytest.run()
+
+
+# %%
+@pytest.mark.parametrize("DF, dims", dac.zipped_ALL)
+def test_edit_y_ticklabels_log_minor(DF, dims):
+    DA = plst.DataAnalysis(data=DF, dims=dims, verbose=False)
+    DA.plot().edit_y_scale_log(base=2).edit_y_ticklabels_log_minor(
+        subs=[2, 3, 5, 7],
+    )
+    if __name__ == "__main__":
+        plt.show()
+    else:
+        plt.close("all")
+
+
+if __name__ == "__main__":
+    ipytest.run()
+
+
+# %%
+labels_zip = [
+    ["sdfsfd", "dddd"],
+    ["sdfsfd", "dddd"],
+    ["sdfsfd", "dddd"],
+    ["sdfsfd", "dddd"],
+]
+zipped_tips = dac.add_zip_column(dac.zipped_noempty_tips, labels_zip)
+
+
+@pytest.mark.parametrize("DF, dims, labels", zipped_tips)
+def test_edit_x_ticklabels_exchange(DF, dims, labels):
+    DA = plst.DataAnalysis(data=DF, dims=dims, verbose=False)
+    DA.plot().edit_x_ticklabels_exchange(
+        labels=labels,
+        labels_lowest_row=[l.upper() for l in labels],
+    )
+    if __name__ == "__main__":
+        plt.show()
+    else:
+        plt.close("all")
+
+
+if __name__ == "__main__":
+    ipytest.run()
+
+
+# %%
+@pytest.mark.parametrize("DF, dims", dac.zipped_ALL)
+def test_edit_x_ticklabels_exchange(DF, dims):
+    DA = plst.DataAnalysis(data=DF, dims=dims, verbose=False)
+    DA.plot().edit_x_ticklabels_rotate(
+        rotation=75,
+        ha="center",
+        # va="top",
+        pad=0.1,
+    )
+    if __name__ == "__main__":
+        plt.show()
+    else:
+        plt.close("all")
+
+
+if __name__ == "__main__":
+    ipytest.run()
+
+# %%
[…]
