### 2023-11-28 ae333c6
added plotting to docs
