+++ pyproject.toml
@@ -74,10 +74,10 @@ requires-python = ">=3.11"
 dependencies = [
     #* Core
     "numpy",
-    "pandas==1.5.3", # !! pingouin Not working with pandas 2.0 yet
+    "pandas==1.5.3", #!! pingouin Not working with pandas 2.0 yet
     # * Plotting
     "matplotlib",
-    "seaborn",
+    "seaborn<=0.12.2", #!! 0.13 has issues with hue
     #* Statistics
     "scipy",
     # "statannot",   #' Superseded by statannotations
