### 2023-11-29 96c15e5
fixed legend position. Added documentation. More
