+++ src/plotastic/plotting/plotedits.py
@@ -1,10 +1,13 @@
 #
-# %% imports
+# %%
+# == imports ===========================================================
 from os import remove
 from typing import TYPE_CHECKING, Callable
 
 import matplotlib as mpl
 import matplotlib.pyplot as plt
+
+import seaborn as sns
 
 # import pyperclip
 
@@ -21,7 +24,7 @@ if TYPE_CHECKING:
 
 class PlotEdits(PlotTool):
     #
-    # == __init__ .......................................................................
+    # == __init__ ======================================================
 
     def __init__(self, **dataframetool_kws):
         super().__init__(**dataframetool_kws)
@@ -29,14 +32,7 @@ class PlotEdits(PlotTool):
         self._edit_y_scalechanged = False
 
     #
-    # == EDIT .........................................................................
-
-    #
-    #' Shapes & Sizes ........................................#
-    # TODO: Until now, stick with the arguments supplied to self.subplots
-
-    #
-    #' Titles of axes .........................................#
+    # == Titles of axes ================================================
 
     @staticmethod
     def _standard_axtitle(
@@ -202,7 +198,7 @@ class PlotEdits(PlotTool):
         return self
 
     #
-    #' Scale of x- & y-axis  ..................................#
+    # == Scale of x- & y-axis ==========================================
 
     def edit_y_scale_log(
         self, base=10, nonpositive="clip", subs=[2, 3, 4, 5, 6, 7, 8, 9]
@@ -211,13 +207,15 @@ class PlotEdits(PlotTool):
             ax.set_yscale(
                 value="log",  #' "symlog", "linear", "logit", ...
                 base=base,  #' Base of the logarithm
-                nonpositive=nonpositive,  #' "mask": masked as invalid, "clip": clipped to a very small positive number
+                #' "mask": masked as invalid, "clip": clipped to a very small positive number
+                nonpositive=nonpositive,
                 subs=subs,  #' Where to place subticks between major ticks ! not working
             )
 
             # ax.yaxis.sety_ticks()
 
-        ### Set the scaled flag to True, to warn user that annotations should be called after NOT before
+        ### Set the scaled flag to True, to warn user
+        #' that annotations should be called after NOT before
         self._edit_y_scalechanged = True
 
         return self
@@ -229,13 +227,14 @@ class PlotEdits(PlotTool):
             ax.set_xscale(
                 value="log",  #' "symlog", "linear", "logit", ...
                 base=base,  #' Base of the logarithm
-                nonpositive=nonpositive,  #' "mask": masked as invalid, "clip": clipped to a very small positive number
+                #' "mask": masked as invalid, "clip": clipped to a very small positive number
+                nonpositive=nonpositive,
                 subs=subs,  #' Where to place subticks between major ticks ! not working
             )
         return self
 
     #
-    #' Ticks and their Labels .................................#
+    # == Ticks and their Labels ========================================
 
     def edit_y_ticklabel_percentage(
         self, decimals_major: int = 0, decimals_minor: int = 0
@@ -251,10 +250,13 @@ class PlotEdits(PlotTool):
         return self
 
     def edit_y_ticklabels_log_minor(self, subs: list = [2, 3, 5, 7]):
-        """Displays minor ticklabels for log-scales. Only shows those ticks whose rounded mantissa (the digits from a float) is in subs
+        """Displays minor ticklabels for log-scales. Only shows those
+        ticks whose rounded mantissa (the digits from a float) is in
+        subs
 
         Args:
-            subs (list, optional): Mantissas (the digits from a float). Defaults to [2, 3, 5, 7].
[…]
