+++ IMAGES/EXAMPLE GALLERY/example_gallery.py
@@ -1,1 +1,6 @@
 """Script that draws all the images in the EXAMPLE GALLERY folder."""
+
+
+### Connect points by subject
+
+### 
