+++ IMAGES/EXAMPLE GALLERY/example_gallery.py
@@ -3,4 +3,6 @@
 
 ### Connect points by subject
 
-### 
+### Custimized boxplot
+
+### Violinplots
