### 2023-11-30 d32a2a6
added instructions how to test this software
