### 2023-11-30 4a46de3
added note that 0.1.1 isn't part of pypi yet
