+++ README.md
@@ -30,53 +30,29 @@ pip install git+https://github.com/markur4/plotastic.git
 *(click to unfold)*
 
 [//]:<------------------------------------------------------------------------->
-<details><summary> 🤔<b><i> Why use plotastic?  </i></b> </summary>
+<details><summary> 🤔<b><i> Summary  </i></b> </summary>
 <blockquote>
 <hr>
 
-#### Statistics made Posssible for EVERYONE:
-- Well-known and intuitive parameters used in `seaborn` (***x***,
-   ***y***, ***hue***, ***row***, ***col***) are 'translated' into terms
-  used for inferential statistics (*between*, *within*, *dv*, etc.) 
-  - **-> *If you know how to plot with seaborn, you can apply basic
-    statistical analyses!***
-- No need need to retype the same arguments of column names into all
-  different tests!
 
-#### Optimized Plotting with `matplotlib`:
-- Make multi-layered pre-configured plots in just one line!
-
-#### Sturdy:
-- plotastic doesn't re-invent the wheel: It's focused on using well
-  established classes, functions and libraries (`pd.DataFrame`,
-  `plt.subplots`, `sns.catplot`, pingouin, statannotations, etc). It's
-  just a wrapper that makes it easier to use them together!
-- plotastic provides feedback on how each step of data import,
-  transformation, formatting or categorization has affected your table,
-  giving beginners the confidence of knowing what they're doing!
-  
-#### Controllable:
-- plotastic outputs common matplotlib figures (`ax`, `fig`). You can
-  modify them like any other!
-- User keyword arguments are passed through plotastic to `seaborn` and
-  `pingouin`, so you can use all their options!
+`plotastic` addresses the challenges of transitioning from exploratory
+data analysis to hypothesis testing in Python's data science ecosystem.
+Bridging the gap between `seaborn` and `pingouin`, this library offers a
+unified environment for plotting and statistical analysis. It simplifies
+the workflow with a user-friendly syntax and seamless integration with
+familiar `seaborn` parameters (y, x, hue, row, col). Inspired by
+`seaborn`'s consistency, `plotastic` utilizes a `DataAnalysis` object to
+intelligently pass parameters to `pingouin` statistical functions. The
+library systematically groups the data according to the needs of
+statistical tests and plots, conducts visualisation, analyses and
+supports extensive customization options. In essence, `plotastic`
+establishes a protocol for configuring statical analyses through
+plotting parameters. This approach streamlines the process, translating
+`seaborn` parameters into statistical terms, providing researchers and
+data scientists with a cohesive and user-friendly solution in python.!
 
 
-<!-- #### Reviewable:
-- We provide snippets that demonstrate of what just happened under the hood, so you can
-  backcheck and thoroughly document your work! -->
-
-[//]:<-- end of 🤔 Why use plotastic? ----------------------------------------->
-</blockquote>
-</details>
-
-
-
-
-[//]:<------------------------------------------------------------------------->
[…]
