+++ src/plotastic/plotting/rc.py
@@ -13,22 +13,24 @@ FONTSIZE = 10
 # %%
 # == STYLE PAPER =======================================================
 PAPER = {
-    # == Figure
+    ### Figure
     "figure.dpi": 200,  #' Displaying figures doesn't need as much dpi as saving them
     "figure.figsize": (3, 3),  #' default is way too big
-    "figure.facecolor": "grey",  #' it's easier on the eyes
-    # == Savefig
+    # "figure.facecolor": "gray",  #' it's easier on the eyes
+    ### Savefig
     "savefig.dpi": 300,  #' Saving figures needs more dpi
     "savefig.format": "pdf",
-    "savefig.transparent": True,
-    # == Font
+    # "savefig.transparent": True,
+    "savefig.facecolor": "white",
+    "axes.facecolor": "white",
+    ### Font
     "font.family": "sans-serif",
     "font.sans-serif": "Arial Narrow",
     "font.size": FONTSIZE,
     "font.weight": "bold",
-    # == Lines
+    # ## Lines
     "lines.linewidth": 0.75,
-    # == Axes
+    # ## Axes
     "axes.spines.right": True,  #' requires argument despine=False
     "axes.spines.top": True,
     "axes.linewidth": 0.75,
@@ -36,11 +38,11 @@ PAPER = {
     "axes.titleweight": "bold",
     "axes.titlepad": 5,
     "axes.labelsize": FONTSIZE,  #' fontsize of the x any y labels
-    # == Grid
-    "axes.grid": True,
+    # ## Grid
+    # "axes.grid": True,
     "grid.linestyle": "-",
     "grid.linewidth": 0.5,
-    # == Ticks
+    # ## Ticks
     "ytick.left": True,
     "xtick.labelsize": FONTSIZE - 1,
     "ytick.labelsize": FONTSIZE - 1,
@@ -52,7 +54,7 @@ PAPER = {
     "ytick.minor.size": 2,
     "xtick.major.size": 2.5,
     "xtick.minor.size": 2,
-    # == Legend
+    # ## Legend
     "legend.fancybox": False,  #' use rounded box for legend
     "legend.title_fontsize": FONTSIZE,
     "legend.fontsize": FONTSIZE,
