+++ IMAGES/EXAMPLE GALLERY/example_gallery.py
@@ -1,8 +1,0 @@
-"""Script that draws all the images in the EXAMPLE GALLERY folder."""
-
-
-### Connect points by subject
-
-### Custimized boxplot
-
-### Violinplots
