### 2023-12-01 ffabe2d
example gallery: diamonds
