+++ README.md
@@ -37,7 +37,7 @@ pip install git+https://github.com/markur4/plotastic.git
   <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/tips1.png" width="350px" alt="logo">
   <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/iris1.png" width="400px" alt="logo">
   <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/cars1.png" alt="logo">
-  <a href="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/diamonds.ipynb">
+  <a href="https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/diamonds.ipynb">
   <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/diamonds1.png" alt="logo"> 
   <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/diamonds2.png" alt="logo">
   </a>
