+++ README.md
@@ -33,14 +33,46 @@ pip install git+https://github.com/markur4/plotastic.git
 <hr>
 
 <h1 align="center">
-  <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/qpcr1.png" alt="logo">
-  <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/tips1.png" width="350px" alt="logo">
-  <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/iris1.png" width="400px" alt="logo">
-  <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/cars1.png" alt="logo">
-  <a href="https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/diamonds.ipynb">
-  <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/diamonds1.png" alt="logo"> 
-  <img src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/diamonds2.png" alt="logo">
+
+  <a href=https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/qpcr.ipynb>
+  <img
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/qpcr1.png"
+  alt="qpcr1">
   </a>
+  
+  <a href=https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/tips.ipynb> 
+  <img
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/tips1.png"
+  width="350px" alt="tips1">
+  </a>
+  
+ <a href=https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/iris.ipynb> 
+  <img
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/iris1.png"
+  width="400px" alt="iris1">
+  </a>
+  
+   <a href="https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/attention.ipynb">
+  <img
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/attention1.png"
+  alt="attention"> 
+  </a
+ 
+  <a href="https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/attention.ipynb"> 
+  <img
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/cars1.png"
+  alt="cars1">
+  </a> 
+  
+  <a href="https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/diamonds.ipynb">
+  <img
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/diamonds1.png"
+  alt="diamonds1">  
+  <img
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/diamonds2.png"
+  alt="diamonds2"> 
+  </a>
+  
 </h1>
 
 
