### 2023-12-03 d2152c4
moved example_gallery to EXAMPLES
