+++ README.md
@@ -34,42 +34,48 @@ pip install git+https://github.com/markur4/plotastic.git
 
 <h1 align="center">
 
-  <a href=https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/qpcr.ipynb>
+  <a href=https://github.com/markur4/plotastic/blob/main/EXAMPLES/qpcr.ipynb>
   <img
-  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/qpcr1.png"
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/EXAMPLES/qpcr1.png"
+  alt="qpcr1">
+  </a 
+
+  <a href=https://github.com/markur4/plotastic/blob/main/EXAMPLES/qpcr.ipynb>
+  <img
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/EXAMPLES/qpcr1.png"
   alt="qpcr1">
   </a>
   
-  <a href=https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/tips.ipynb> 
+  <a href=https://github.com/markur4/plotastic/blob/main/EXAMPLES/tips.ipynb> 
   <img
-  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/tips1.png"
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/EXAMPLES/tips1.png"
   width="350px" alt="tips1">
   </a>
   
- <a href=https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/iris.ipynb> 
+ <a href=https://github.com/markur4/plotastic/blob/main/EXAMPLES/iris.ipynb> 
   <img
-  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/iris1.png"
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/EXAMPLES/iris1.png"
   width="400px" alt="iris1">
   </a>
   
-   <a href="https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/attention.ipynb">
+   <a href="https://github.com/markur4/plotastic/blob/main/EXAMPLES/attention.ipynb">
   <img
-  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/attention1.png"
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/EXAMPLES/attention1.png"
   alt="attention"> 
-  </a
+  </a>
  
-  <a href="https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/cars.ipynb"> 
+  <a href="https://github.com/markur4/plotastic/blob/main/EXAMPLES/cars.ipynb"> 
   <img
-  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/cars1.png"
+  src="https://raw.githubusercontent.com/markur4/plotastic/main/EXAMPLES/cars1.png"
   alt="cars1">
   </a> 
   
-  <a href="https://github.com/markur4/plotastic/blob/main/IMAGES/EXAMPLE_GALLERY/diamonds.ipynb">
+  <a href="https://github.com/markur4/plotastic/blob/main/EXAMPLES/diamonds.ipynb">
   <img
-  src="https://raw.githubusercontent.com/markur4/plotastic/main/IMAGES/EXAMPLE_GALLERY/diamonds1.png"
[…]
