### 2023-12-03 acf9847
added tips to README.md
