### 2023-12-05 a172491
changed order of factors_all to row,col,x,hue
