### 2023-12-06 bbcc20e
updated path to Fig 1 after moved to EXAMPLES
