### 2023-12-07 6a1605d
began implementing dims_hierarchical, subjects,etc
