+++ tests/_multiplot_test.py
@@ -7,7 +7,7 @@ import matplotlib.pyplot as plt
 
 import plotastic as plst
 
-import _DA_configs as dac
+import DA_configs as dac
 
 
 # %%
