+++ tests/_rc_test.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 
 import plotastic as plst
 
-import _DA_configs as dac
+import DA_configs as dac
 
 # %%
 
