+++ tests/_save_test.py
@@ -14,7 +14,7 @@ import pytest
 import plotastic as plst
 
 
-import _DA_configs as dac
+import DA_configs as dac
 
 
 # %% testfigure
