+++ src/plotastic/dimensions/dataframetool.py
@@ -19,7 +19,9 @@ import matplotlib.pyplot as plt
 
 # import markurutils as ut
 import plotastic.utils.utils as ut
-from plotastic.dimensions.dimsandlevels import DimsAndLevels
+
+# from plotastic.dimensions.dimsandlevels import DimsAndLevels
+from plotastic.dimensions.hierarchical_dims import HierarchicalDims
 
 from typing import Tuple
 
@@ -29,18 +31,15 @@ if TYPE_CHECKING:
 # %% Class DataFrameTool
 
 
-class DataFrameTool(DimsAndLevels):
+class DataFrameTool(HierarchicalDims):
     # ==
     # == __init__ ======================================================
 
     def __init__(
         self,
         levels: list[tuple[str]] = None,
-        # levels_ignore: list[tuple[str]] = None,
-        subject: str = None,
-        # transform=None
-        verbose=False,
-        **dims_data_kws,
+        verbose=True,
+        **kws,
     ):
         """Adds pandas DataFrame related tools.
 
@@ -48,29 +47,16 @@ class DataFrameTool(DimsAndLevels):
                 with the dataframe and columns will be set to ordered categorical type
                 automatically, Defaults to None.
         :type levels: list[tuple[str]], optional
-        :param subject: Column defining the subject, which connects samples as dependent
-            (patient, donor, date etc.). If specified, paired tests will be prioritized,
-            defaults to None.
-        :type subject: str, optional
-        :param verbose: Warns User of empty groups, defaults to False
+        :param verbose: Run object in verbose mode, which means it'll
+            print a lot of feedback, Defaults to True.
         :type verbose: bool, optional
         """
 
-        ### Inherit from DimsAndLevels
-        super().__init__(**dims_data_kws)
+        ### Inherit
+        super().__init__(**kws)
 
-        ### Attributes
-        #' to make columns categorical (and exclude levels, if unspecified)
-        self.user_levels = levels
-        #' for paired analysis
-        self.subject = subject
-        if not subject is None:
-            assert (
-                subject in self.data.columns
-            ), f"#! Subject '{subject}' not in columns, expected one of {self.data.columns.to_list()}"
-
-        # ### Initialize a dataframe that contains  groups for every combination of factor levels (with dv = N)
-        # self.data_allgroups = self.data_ensure_allgroups()
+        ### Verbose mode
+        self.verbose = verbose
 
         ### Transformations
         self.transformed = False
@@ -78,22 +64,15 @@ class DataFrameTool(DimsAndLevels):
         #' Store y so we can reset it after transformations
         self._y_untransformed = self.dims.y
 
-        ### Check for empties or missing group combinations
-        self.factors_always_together = None
-        if verbose:
-            self.data_check_integrity()
-
         ### Make Categorical
+        #' Levels are excluded levels, if unspecified
+        self.user_levels = levels  # TODO: Rework this
         if levels:
             self._check_inputlevels_with_data(
                 input_lvls=levels, verbose=verbose
             )
             self.input_levels = levels
             _ = self.data_categorize(verbose=verbose)
-
-        # if levels_ignore:
-        #     self.check_inputlevels_with_data(input_lvls=levels_ignore, verbose=verbose)
-
 
     # ==
     # ==  Make Levels Categorical ======================================
@@ -382,404 +361,6 @@ class DataFrameTool(DimsAndLevels):
         if verbose:
             ut.pp(df)
         return df
-
-    # ==
-    # == COUNT GROUPS AND SAMPLESIZE ===================================
-
-    def data_get_samplesizes(self) -> pd.Series:
-        """Returns a DataFrame with samplesizes per group/facet"""
-        #' Get samplesize per group
-        samplesize_df = (
-            self.data.groupby(self.factors_all)
[…]
