+++ CHANGELOGS.md
@@ -32,3 +32,4 @@
 ### Available on:
 - github
 - pypi
+

[…]
