### 2023-12-13 3c33c5e
minor bibliography change
