### 2023-12-13 0a6789b
removed self.verbose attribute, define per method
