### 2023-12-13 44130e8
verbose is now a switch in DataAnalysis, not attr
