+++ devtools/tomltovenv/__main__.py
@@ -1,251 +1,0 @@
-#
-# %%
-
-import subprocess
-import os
-import argparse
-
-# !! This must be the only place where rc is imported, otherwise user
-# !! arguments will be overwritten if called again later
-from rc import (
-    PROJECT_ROOT,
-    VENV_NAME,
-    REDO_REQUIREMENTS_TXT,
-    EDITABLE,
-    INSTALL_DEVTOOLS,
-)
-
-# %%
-
-# == INIT Parsers ==============================================================
-parser = argparse.ArgumentParser(
-    description="Reads your ``pyproject.toml`` inside project root and sets up a virtual environment including dependencies, devtools, editable install, etc.",
-    add_help=True,
-)
-
-# ==============================================================================
-# == Add Arguments to Parser ===================================================
-# fmt: off
-
-### Reused args
-kws = dict(
-    default=None,  
-    required=False,
-    action="store_true",
-)
-
-### Optional
-parser.add_argument(
-    "-p", "--project-root", 
-    help="Path to project root. Default: %(default)s",
-    type=str, #' Cast type when passing
-    default=".",  #' Ok, since we assert for presence of pyproject.toml
-)
-parser.add_argument(
-    "-v", "--venv-name",
-    help="Name of virtual environment. Default: %(default)s",
-    type=str,
-    default="venv",
-)
-
-### Flags
-parser.add_argument(
-    "-r", "--redo-requirements",
-    help="!! Use Re-solve environment and (over)write requirements.txt. Default: %(default)s",
-    **kws,
-)
-parser.add_argument(
-    "-t", "--test-dir",
-    help="Specify directory within project root to find tests. If not specified, no tests will be run. If only -t is passed without directory, will use 'tests' as values. Requires --install-devtools. Default: %(default)s",
-    type=str,
-    nargs="?",  #' Make optional
-    const=True,  #' If only -t is passed without directory, set to True
-    default=False, # !! overriden later to "tests"
-)
-
-### --user-mode switches off Features
-#' User mode: Mimic a setup of a user after pip install <package>
-#' User mode: Switches off devtools, editable
-parser.add_argument(
-    "-u", "--user-mode",
-    help="Executes standard install of the project (pip install <project>), not editable and without devtools. Default: %(default)s",
-    **kws,
-)
-#' Switch back on despite --user-mode
-parser.add_argument(
-    "-e", "--editable",
-    help="By default, %(prog)s installs project in editable mode, but when switched off by --user-mode, use --editable to switch it back on",
-    **kws,
-)
-parser.add_argument(
-    "-d", "--install-devtools",
-    help="By default, %(prog)s installs developer tools, but when switched off by --user-mode, use --editable to switch it back on",
-    **kws,
-)
-# fmt: on
-
-
-# ==============================================================================
-# == Redirect user input into settings =========================================
-#' Values from rc.py are capitalized: PROJECT_ROOT, VENV_NAME, etc.
-
-### Parse Args
-args = parser.parse_args()
-
-### Paths
-# if args.project_root:  #' -p, string, default to "."
-PROJECT_ROOT = args.project_root
-
-# if args.venv_name:  #' -v, string, default to "venv"
-VENV_NAME = args.venv_name
-
-### Flags
-if not args.redo_requirements is None:  #' -r
-    REDO_REQUIREMENTS_TXT = args.redo_requirements
-
-if args.test_dir is True:
-    test_dir = "tests" #' -t
-else:
[…]
