+++ devtools/tomltovenv/utils_ttv.py
@@ -1,206 +1,0 @@
-"""Utilities for environment setup
-1. Install a Python globally for your OS with correct version
-2. Use global Python to call `venv.create` and make a virtual environment for
-   each project 
-3. Switch Jupyter Kernel to venv
-4. Install Project (`pip install -e .`)
-5. Freeze requirements (`pip freeze > requirements.txt`)
-6. Install requirements for development (`pip install .[dev]`)
-"""
-
-
-# %%
-### Builtin imports
-import subprocess
-import shlex
-import os
-import venv
-
-### Third party imports, requires prepare_venv.py to be run
-import toml
-from icecream import ic  #' For Testing these functions
-
-# from env_config import *
-
-
-### For testing purposes, define location of pyproject.toml
-if __name__ == "__main__":
-    project_root = ".."
-    pyproject_toml = os.path.join(project_root, "pyproject.toml")
-
-
-# %%
-def run_cmd(cmd: str, print_cmd=True, ret_p=False, **kwargs):
-    """Run a command in the shell, e.g. `brew install python@3.11`"""
-
-    ### Define default kwargs and override with user kwargs
-    kws = dict(
-        shell=False,  # !! Avoid shell, defaults to /bin/sh
-    )
-    kws.update(kwargs)
-
-    ### Print command
-    if print_cmd:
-        print("\n$", cmd)
-
-    ### Split command into list
-    cmd: str = shlex.split(cmd)
-
-    ### Run command
-    #' .Popen doesn't output files as easily
-    p = subprocess.run(cmd)
-
-    if ret_p:
-        return p
-
-
-if __name__ == "__main__":
-    c = "ping -c 4 google.com"
-    # c = "echo $PATH"
-    # c = "pip install pyperclip"
-    run_cmd(c)
-
-    # %%
-    # ! pip uninstall pyperclip -y
-    c = "pip install pyperclip"
-    run_cmd(c)
-
-
-# %%
-def check_python(toml_f: str) -> None:
-    print("We need Python Version: ", get_python_version(toml_f))
-    run_cmd("python --version")
-    run_cmd("which python")
-
-
-def check_path():
-    print("Checking PATH:")
-    PATH = subprocess.check_output(["echo $PATH"], shell=True)
-    PATH_list = PATH.decode("utf-8").split(":")
-    for path in PATH_list:
-        print(path)
-    print(len(PATH_list))
-
-
-# %%
-def create_venv(project_root: str, env_name: str) -> None:
-    """Create a virtual environment for the project"""
-    if os.path.isdir(os.path.join(project_root, env_name)):
-        print(f"'{env_name}' will be overwritten!")
-    print(f">>> venv.create({env_name}, clear=True, with_pip=True)")
-    venv.create(
-        env_dir=os.path.join(project_root, env_name),
-        clear=True,  #' Clear the environment directory before creating it
-        with_pip=True,
-    )
-
-
-def make_python_path(project_root: str, env_name: str) -> str:
-    """Get the path to the python interpreter of the virtual environment"""
-    return os.path.join(project_root, env_name, "bin", "python")
-
-
-# %%
-def install(packages: list, python: str = None):
-    """Install packages"""
-    python = "python" if python is None else python
-
-    packages_str = " ".join(packages)
-    run_cmd(f"{python} -m pip install {packages_str}")
-
-
-def install_deps(
-    packages: list = None, requirements_txt: str = None, python: str = None
-) -> None:
-    """Install dependencies from requirements.txt, if provided, else
-    from pyproject.toml
-    """
-    python = "python" if python is None else python
-
-    if not requirements_txt is None:
[…]
