+++ devtools/tomltovenv/README.rst
@@ -1,15 +1,0 @@
-``tomltovenv``
-==============
-
-**A tool that reads your ``pyproject.toml`` and ...**
-
-*  ... creates a virtualenv in the project root
-*  ... installs dependencies
-*  ... freezes the environment to a ``requirements.txt``
-*  ... installs developer tools
-*  ... installs the project in editable mode
-*  ... runs tests
-
-**... all in one command.**
-
-

[…]
