### 2024-01-27 4b74f64
assert subject, type hints, pillow version
