+++ tests/DA_configs2.py
@@ -1,0 +1,9 @@
+
+### They are all non-empty
+StatTestCases = [
+    "all",
+    "paired", 
+    "unpaired",
+    "unpaired",
+    "parametric_paired",
+]
