+++ requirements.txt
@@ -35,7 +35,7 @@ pandas-flavor==0.6.0
 parso==0.8.3
 patsy==0.5.3
 pexpect==4.8.0
-Pillow==10.1.0
+Pillow>=10.2.0
 pingouin==0.5.3
 platformdirs==4.0.0
 prompt-toolkit==3.0.41
