### 2024-01-27 ef72c85
added setupvenv to devtools, checked pillow
