+++ requirements.txt
@@ -1,66 +1,58 @@
 appnope==0.1.3
 asttokens==2.4.1
-certifi==2023.11.17
-charset-normalizer==3.3.2
 colorama==0.4.6
 colour==0.1.5
-comm==0.2.0
+comm==0.2.1
 contourpy==1.2.0
 cycler==0.12.1
 debugpy==1.8.0
 decorator==5.1.1
 et-xmlfile==1.1.0
 executing==2.0.1
-fonttools==4.45.0
+fonttools==4.47.2
 icecream==2.1.3
-idna==3.4
-ipykernel==6.26.0
+ipykernel==6.29.0
 ipynbname==2023.2.0.0
-ipython==8.17.2
+ipython==8.20.0
 jedi==0.19.1
 joblib==1.3.2
 jupyter_client==8.6.0
-jupyter_core==5.5.0
+jupyter_core==5.7.1
 kiwisolver==1.4.5
-littleutils==0.2.2
 matplotlib==3.8.2
 matplotlib-inline==0.1.6
-nest-asyncio==1.5.8
-numpy==1.26.2
+nest-asyncio==1.6.0
+numpy==1.26.3
 openpyxl==3.1.2
-outdated==0.2.2
 packaging==23.2
 pandas==1.5.3
 pandas-flavor==0.6.0
 parso==0.8.3
-patsy==0.5.3
-pexpect==4.8.0
-Pillow>=10.2.0
-pingouin==0.5.3
-platformdirs==4.0.0
-prompt-toolkit==3.0.41
-psutil==5.9.6
+patsy==0.5.6
+pexpect==4.9.0
+pillow==10.2.0
+pingouin==0.5.4
+platformdirs==4.1.0
+prompt-toolkit==3.0.43
+psutil==5.9.8
 ptyprocess==0.7.0
 pure-eval==0.2.2
-Pygments==2.17.1
+Pygments==2.17.2
 pyparsing==3.1.1
 python-dateutil==2.8.2
 pytz==2023.3.post1
-pyzmq==25.1.1
-requests==2.31.0
-scikit-learn==1.3.2
-scipy==1.11.4
+pyzmq==25.1.2
+scikit-learn==1.4.0
+scipy==1.12.0
 seaborn==0.11.2
 six==1.16.0
 stack-data==0.6.3
 statannotations==0.6.0
-statsmodels==0.14.0
+statsmodels==0.14.1
 tabulate==0.9.0
 threadpoolctl==3.2.0
-toml==0.10.2
-tornado==6.3.3
-traitlets==5.13.0
-urllib3==2.1.0
-wcwidth==0.2.10
-xarray==2023.11.0
+tornado==6.4
+traitlets==5.14.1
+wcwidth==0.2.13
+xarray==2024.1.1
 XlsxWriter==3.1.9
