### 2024-01-29 #6304 @editorialbot
[REVIEW]: plotastic: Bridging Plotting and Statistics in Python
**Submitting author:** @markur4 (<a href="http://orcid.org/0009-0009-7292-7714">Martin Kuric</a>)
**Repository:** https://github.com/markur4/plotastic
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.1
