### 2024-02-07 markur4/plotastic#1 @SunnyXu
Manuscript improvement
Please pay attention to some points in the manuscript.
(1) There is not supposed to be a period mark before the citation on line 26, and the citation on line 37 with the front bracket placed at a wrong position.
(2) Figure 1 is not cited in the content of the manuscript.
