### 2024-02-07 markur4/plotastic#2 @SunnyXu
PYPI
(1) Please note that the figures/plots on the PYPI page (https://pypi.org/project/plotastic/) are not shown.
(2) The package indeed works with Python 3.11. However, is this Python package only applicable to 3.11 or did you perform it with Python 3.11 only?
