### 2024-02-07 markur4/plotastic#3 @SunnyXu
Data of Example Gallery in the section of Documentations
It would be good to create a folder to store all the data used in the Example Gallery (iii-vi) or at least give a guide about how to download them to use and reproduce your examples.
