### 2024-02-07 markur4/plotastic#4 @SunnyXu
Some parts of the Documentation are empty
Please either remove or complete them.
1 Example Gallery– i Quick Example: FMRI: it shows 404
https://github.com/markur4/plotastic/blob/main/HOW_TO_USE/quick_example_fmri.ipynb
2 Data- i Transform Y: no link
3 Plotting- iii Editing Figures: no content
