### 2024-02-07 markur4/plotastic#5 @SunnyXu
Quick Example 
In the section of Quick Example on GitHub README:
Please add “import matplotlib.pyplot as plt” in the front.
