### 2024-02-07 markur4/plotastic#6 @SunnyXu
Fix the point of “Community guidelines” in the review checklist
Please add a clear statement of Community Guidelines refer to:
https://github.com/openjournals/joss-reviews/issues/6304#issuecomment-1932772435
