### 2024-02-07 #6304 @SunnyXu
@rkurchin @markur4
plotastic is a solid Python package bridging the gap between seaborn and pingouin to offer a unified environment for plotting and statistical analysis. It is good that there are the testing section and codecov on GitHub. Please see my comments (linked to issues) below to improve. Please let me know if anything is unclear.
1. Manuscript improvement
(https://github.com/markur4/plotastic/issues/1)
2. PYPI
(https://github.com/markur4/plotastic/issues/2)
3. Data of Example Gallery in the section of Documentations (https://github.com/markur4/plotastic/issues/3)
4. Some parts of the Documentation are empty (https://github.com/markur4/plotastic/issues/4)
5. Quick Example
(https://github.com/markur4/plotastic/issues/5)
6. fix the point of “Community guidelines” in my review checklist (https://github.com/markur4/plotastic/issues/6)
