### 2024-02-15 markur4/plotastic#1 @markur4
Thank you for pointing out these points! I have made these edits:
(1)
  -  Former line 26: The period before (Waskom, 2021) has been removed.
  -  Former line 37: The citations have been corrected.
(2) Fig. 1 has been referenced after the first sentence of the 'example section'
