### 2024-02-15 markur4/plotastic#2 @markur4
Thanks for pointing these out!
(1) For tables, I've decided to exclude them from PyPI, since they weren't displaying correctly. For figures/plots, I decided to exclude them at early versions, because at that time, I was changing the image directories and the links would keep breaking. I plan on displaying figures/plots in PyPI in later versions.

(2) plotastic runs on python==3.11, since I haven't tested older versions. I could run plotastic on older versions to see which is the minimum version.
