### 2024-02-15 markur4/plotastic#3 @markur4
Good point! All the images in the image gallery are clickable, linking to the .ipynb that shows the full code. These notebooks retrieve dataframes by `sns.load_dataset` & `plst.load_dataset`.

I made these revisions to make reproduction more accessible:
- Added "Click on Images for Code" at the top of Example Gallery of the README
- Added "Raw Data" and "Original Source" underneath the header of each example with explicit links to the seaborn databases. For databases implemented by plotastic, it leads to the example_data folder found in the plotastic src
