### 2024-02-15 markur4/plotastic#4 @markur4
I appreciate the detailed review! I made these corrections:
(1) Corrected the link to fmri
(2) Removed Transform Y (will be added later)
(3) Removed Editing Figures (will be added later)
