### 2024-02-15 42c4e9f
fixed links, added plt import, removed empty docs
