### 2024-02-15 markur4/plotastic#6 @markur4
I have added a "Community Guidelines" Section to the Readme. Please let me know, if it's ok!
