### 2024-02-16 markur4/plotastic#1 @SunnyXu
Hi, I have just noticed that there is another period mark before the citation on line 26.
