### 2024-02-28 #6304 @gmrandazzo
@markur4 and @rkurchin

Thanks for the chance to review this paper. Here are my comments:

1) The summary is OK; however, there needs to be a reason why it would be good to use plastic instead of seaborn+pingouin. This important statement is written in lines 54, 55, and 56 of the manuscript. Please report them also in the summary. The rest of the manuscripts seem fine to me.

2) Installation is fine. However, this package works only for Python >=3.11, and I need clarification on why older Python, i.e., 3.9, is not OK. If there is any specific reason, please state this in the installation document, specifying also the minimum Python version. Can you please fix that?

from pyproject.toml
```
[...]
#' Specify version only if concrete incompatibilities exist
requires-python = ">=3.11"
```

Is there any specific incompatibility? If yes, state this. If not, please fix that.

3) I have opened a shell. I copied and pasted the script you presented in the paper, and the Python shell became unusable. I'm not able to see the final plot. Can you check what is going on?

4) Functional documentation seems poor. Please improve it.

Thanks!
Best regards
