### 2024-02-29 a506217
explained python version
