### 2024-02-29 c7223c4
annotate_pairwise type annotation include str
