+++ src/docs/index.rst
@@ -1,20 +1,0 @@
-.. plotastic documentation master file, created by
-   sphinx-quickstart on Thu Sep  7 20:25:33 2023.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to plotastic's documentation!
-=====================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   modules
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
