+++ src/docs/modules.rst
@@ -1,7 +1,0 @@
-src
-===
-
-.. toctree::
-   :maxdepth: 4
-
-   plotastic
