+++ src/docs/plotastic.dataanalysis.rst
@@ -1,29 +1,0 @@
-plotastic.dataanalysis package
-==============================
-
-Submodules
-----------
-
-plotastic.dataanalysis.annotator module
----------------------------------------
-
-.. automodule:: plotastic.dataanalysis.annotator
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-plotastic.dataanalysis.dataanalysis module
-------------------------------------------
-
-.. automodule:: plotastic.dataanalysis.dataanalysis
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: plotastic.dataanalysis
-   :members:
-   :undoc-members:
-   :show-inheritance:
