+++ src/docs/plotastic.stat.rst
@@ -1,61 +1,0 @@
-plotastic.stat package
-======================
-
-Submodules
-----------
-
-plotastic.stat.assumptions module
----------------------------------
-
-.. automodule:: plotastic.stat.assumptions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-plotastic.stat.bivariate module
--------------------------------
-
-.. automodule:: plotastic.stat.bivariate
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-plotastic.stat.omnibus module
------------------------------
-
-.. automodule:: plotastic.stat.omnibus
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-plotastic.stat.posthoc module
------------------------------
-
-.. automodule:: plotastic.stat.posthoc
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-plotastic.stat.statresults module
----------------------------------
-
-.. automodule:: plotastic.stat.statresults
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-plotastic.stat.stattest module
-------------------------------
-
-.. automodule:: plotastic.stat.stattest
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: plotastic.stat
-   :members:
-   :undoc-members:
-   :show-inheritance:
