### 2024-02-29 #6304 @markur4
@gmrandazzo @rkurchin

Thank you very much for the valuable feedback!

## 1.
That's a good point, I included the rationale in the last sentence of the summary:

**Before:** [...]_This approach streamlines the process, translating `seaborn` parameters into statistical terms, providing researchers and data scientists with a cohesive and user-friendly solution in python._

**After:** [...] _This approach streamlines the process, translating `seaborn` parameters into statistical terms, allowing researchers to focus on correct statistical testing and less about specific syntax and implementations._

## 2.
I haven't tested plotastic on any other version than 3.11. I revised pyproject.toml to reduce confusion:
```
### Python version
#' Lower versions than 3.11 have not been tested
requires-python = ">=3.11"
```
Please let me know if testing on older python versions is required to pass the review.

## 3.
Interesting, I have never executed plotastic in a shell. Here's what I did:
- Made "tests/_paper_test.py" and copy pasted code from paper into
- Executed "tests/_paper_test.py" in vscode interactive mode (venv editable mode): **this worked!**
- opened shell, ran this:
`python _paper_test.py`
**this worked!** (no plot shows up, but shell remains usable)
- Made a fresh venv, installed from github according to readme
- Ran `python _paper_test.py` in both shell and vscode: **this worked, too**
- Opened a python shell, copy pasted code: **this worked!** (but no plot shows up!)
- In the same python shell, I ran:
```
import seaborn as sns
sns.catplot(data=DF, **_dims)
```
- **this worked!** but also did NOT open a plot!

**Summary:**
- Sadly, I can not reproduce your shell getting unresponsive
- Since seaborn code also does not open plots when running from the python shell, I don't think that plotastic should be able to do this, too.
- plotastic was designed to be used with ipython, I would not recommend execution in python shell at all.
- Could you please retry using ipython / jupyter notebooks / vscode interactive mode?

## 4.
The screenshots do not refer to the correct documentation, but preliminary experiments with sphinx. I git removed them completely to prevent confusion.

Please use the documentations and examples found in the README.md as a guide to use plotastic. Let me know if they are clear!

## 5.
Thanks for pointing out the garbage files! I git removed them from src, if you find any more, let me know!
