### 2024-02-29 #6304 @gmrandazzo
@markur4 ok for point 2. State the minimum version in the README in the section "REQUIREMENTS" or something like that. Try to harmonize everything. Remember that someone landing your GitHub page should be able to install the software without somersault.

Then @rkurchin fine to me. Accepted.
