### 2024-03-03 3f21709
removed fragmentaric sentence from summary
