+++ README.md
@@ -15,13 +15,15 @@ black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://gith
 
 
 [//]:<== Installation =========================================================>
-# Installation 📦
+# 📦 Installation
 
+#### Install from PyPi:
 ``` bash
-# Install from PyPi:
 pip install plotastic
+```
 
-# For the latest version (check CHANGELOG.md!):
+#### Install from GitHub: (experimental, check CHANGELOG.md)
+``` bash
 pip install git+https://github.com/markur4/plotastic.git
 ```
 
@@ -31,7 +33,7 @@ pip install git+https://github.com/markur4/plotastic.git
 - seaborn <= 0.12.2 *(later versions reworked hue)*
 
 [//]:<== Information ==========================================================>
-# Example Gallery 📷
+#  📷 Example Gallery
 
 <details><summary> <b><i> (click to unfold)  </i></b> </summary>
 <blockquote>
@@ -92,7 +94,7 @@ pip install git+https://github.com/markur4/plotastic.git
 
 
 [//]:<== Information ==========================================================>
-# About plotastic 📚 
+# 🧑‍🏫 About `plotastic` 
 
 
 [//]:<------------------------------------------------------------------------->
@@ -294,7 +296,7 @@ data scientists with a cohesive and user-friendly solution in python.!
 
 
 [//]:<== Features =============================================================>
-## Features ⚙️
+
 
 
 <details><summary>  ✅ <b><i> Feature List </i></b> </summary>
@@ -704,30 +706,16 @@ classDiagram
 
 
 
-[//]:<=========================================================================>
-## Cite this! ✍🏻
-<details><summary> <i> Please cite the publications of seaborn, pingouin, etc. when using plotastic (click to unfold) </i> </summary>
-<blockquote>
-<hr>
-
-- *Vallat, R. (2018). Pingouin: statistics in Python. Journal of Open
-  Source Software, 3(31), 1026. <https://doi.org/10.21105/joss.01026>*
-- *Waskom, M. et al. (2021). mwaskom/seaborn: v0.11.1 (January 2021).
-  Zenodo. <http://doi.org/10.5281/zenodo.4547176>*
-
-[//]:<end Cite this! ✍🏻 =======================================================>
-</blockquote>
-</details>
 
 
 
 
-# How To Use 📔
+# 📔 How To Use
 
 [//]:<=========================================================================>
 
-## Documentations 💡
-1. Example Gallery
+## Documentations
+1. **Example Gallery**
    1. [Quick Example: FMRI](https://github.com/markur4/plotastic/blob/main/EXAMPLES/fmri.ipynb)
    2. [qPCR (paired, parametric)](https://github.com/markur4/plotastic/blob/main/EXAMPLES/qpcr.ipynb)
    3. [Cars (unpaired, non-parametric)](https://github.com/markur4/plotastic/blob/main/EXAMPLES/cars.ipynb)
@@ -735,16 +723,17 @@ classDiagram
    5. [Attention (paired/mixed, parametric)](https://github.com/markur4/plotastic/blob/main/EXAMPLES/attention.ipynb)
    6. [Iris (unpaired, parametric)](https://github.com/markur4/plotastic/blob/main/EXAMPLES/iris.ipynb)
    7. [Tips (unpaired, parametric)](https://github.com/markur4/plotastic/blob/main/EXAMPLES/tips.ipynb)
-2. Data
+2. **Data**
    <!-- 1. Transform Y -->
[…]
