### 2024-03-03 #6304 @markur4
@rkurchin

Here's what I did:

- Built current repo, uploaded on [PyPi](https://pypi.org/project/plotastic/),
**Version: 0.1.1**
- Cloned repo, zipped it and uploaded on Zenodo,
**DOI: 10.5281/zenodo.10775033**
- Checked that license (GPLv3), ORCIDs and authors are the same as in the archive

Let me know if I missed something!
