### 2024-03-06 #6304 @rkurchin
Apologies that this took a little while!

Editorial comments:
* line 9: I think you can remove the "a" before "syntax" since I usually parse that as a collective noun
* line 12-13: I'm not sure if "analyses" is the plural of "analysis" (and hence an additional object of the verb "conducts") or the British English spelling of the third-person singular of the verb "analyzes" – in any case, the parallelization of the sentence is a bit broken. I would suggest rewriting/splitting it. One possibility: "The library systematically groups the data according to the needs of statistical tests and plots. It can perform these analyses and visualize the results and also supports extensive customization options"
* 22-25: not sure what's going on with the line spacing thing here, but perhaps double-check if there's anything goofy about your source file?
* 28: remove extra period before citation
* 31: missing a "with" before `seaborn`
* figure 1: Why is the whole of the plotting block in parentheses?
* 42: I think the apostrophe after `seaborn` is a ` instead of a '
* 67: "per axes" --> "per axis"
* 71-72: add "a" after "initializes"
* 77, 91: plotastic --> `plotastic`
* 78: remove "in"
* 82: there's a word missing in this sentence...statistical what? Perhaps information? Indicators? Parameters?
* References: [pandas requests](https://pandas.pydata.org/about/citing.html) two particular things be cited when acknowledging it, please replace your citation with those two
