+++ paper.md
@@ -29,24 +29,27 @@ aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
 # aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
+
 # Summary
 
 `plotastic` addresses the challenges of transitioning from exploratory
 data analysis to hypothesis testing in Python's data science ecosystem.
 Bridging the gap between `seaborn` and `pingouin`, this library offers a
 unified environment for plotting and statistical analysis. It simplifies
-the workflow with a user-friendly syntax and seamless integration with
+the workflow with user-friendly syntax and seamless integration with
 familiar `seaborn` parameters (y, x, hue, row, col). Inspired by
 `seaborn`'s consistency, `plotastic` utilizes a `DataAnalysis` object to
-intelligently pass parameters to `pingouin` statistical functions. The
-library systematically groups the data according to the needs of
-statistical tests and plots, conducts visualisation, analyses and
-supports extensive customization options. In essence, `plotastic`
-establishes a protocol for configuring statical analyses through
-plotting parameters. This approach streamlines the process, translating
-`seaborn` parameters into statistical terms, allowing researchers to
-focus on correct statistical testing and less about specific syntax and
+intelligently pass parameters to `pingouin` statistical functions.
+Hence, statistics and plotting are performed on the same set of
+parameters, so that the strength of `seaborn` in visualizing
+multidimensional data is extended onto statistical analysis. In essence,
+`plotastic` translates `seaborn` parameters into statistical terms,
+configures statistical protocols based on intuitive plotting syntax and
+returns a `matplotlib` figure with known customization options and more.
+This approach streamlines data analysis, allowing researchers to focus
+on correct statistical testing and less about specific syntax and
 implementations.
+
 
 
 # Statement of need
@@ -55,24 +58,21 @@ Python's data science ecosystem provides powerful tools for both
 visualization and statistical testing. However, the transition from
 exploratory data analysis to hypothesis testing can be cumbersome,
 requiring users to switch between libraries and adapt to different
-syntaxes.
-
-`seaborn` has become a popular choice for plotting in Python, offering
-an intuitive interface. Its statistical functionality focuses on
-descriptive plots and bootstrapped confidence intervals
-
+syntaxes. `seaborn` has become a popular choice for plotting in Python,
+offering an intuitive interface. Its statistical functionality focuses
+on descriptive plots and bootstrapped confidence intervals
 [@waskomSeabornStatisticalData2021]. The library `pingouin` offers an
 extensive set of statistical tests, but it lacks integration with common
 plotting capabilities [@vallatPingouinStatisticsPython2018].
-`statannotations` integrates statistical testing with plot annotations, but
- uses a complex interface and is limited to pairwise comparisons
+ `statannotations` integrates statistical testing with plot annotations,
+ but uses a complex interface and is limited to pairwise comparisons
  [@charlierTrevismdStatannotationsV02022].
 
 `plotastic` addresses this gap by offering a unified environment for
 plotting and statistical analysis. With an emphasis on user-friendly
-syntax and integration with familiar `seaborn` parameters, it simplifies
-the process for users already comfortable `seaborn`. The library ensures
-a smooth workflow, from data import to hypothesis testing and
+syntax and integration of familiar `seaborn` parameters, it simplifies
+the process for users already comfortable with `seaborn`. The library
+ensures a smooth workflow, from data import to hypothesis testing and
 visualization.
 
 # Example
@@ -169,11 +169,10 @@ of statistical analysis and plotting, leveraging the capabilities of
[…]
