+++ paper.bib
@@ -5,7 +5,6 @@
   year = {2022},
   month = oct,
   doi = {10.5281/ZENODO.7213391},
-  url = {https://zenodo.org/record/7213391},
   urldate = {2023-11-16},
   abstract = {Add scipy's Brunner-Munzel test Fix applying statannotations for non-string group labels (Issue \#65) Get Zenodo DOI},
   copyright = {Open Access},
@@ -24,10 +23,18 @@
   pages = {90--95},
   issn = {1558-366X},
   doi = {10.1109/MCSE.2007.55},
-  url = {https://ieeexplore.ieee.org/document/4160265},
   urldate = {2023-11-15},
   abstract = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting,and publication-quality image generation across user interfaces and operating systems},
   file = {/Users/martinkuric/Zotero/storage/W4FJZDNY/§-hunterMatplotlib2DGraphics2007.pdf;/Users/martinkuric/Zotero/storage/GW3HZZHR/4160265.html}
+}
+
+@inproceedings{mckinneyDataStructuresStatistical2010,
+  title = {Data {{Structures}} for {{Statistical Computing}} in {{Python}}},
+  author = {McKinney, Wes},
+  year = {2010},
+  month = jan,
+  pages = {56--61},
+  doi = {10.25080/Majora-92bf1922-00a}
 }
 
 @article{mckinneyPandasFoundationalPython2011,
@@ -39,6 +46,15 @@
   journal = {Python High Performance Science Computer},
   abstract = {---In this paper we will discuss pandas, a Python library of rich data structures and tools for working with structured data sets common to statistics, finance, social sciences, and many other fields. The library provides integrated, intuitive routines for performing common data manipulations and analysis on such data sets. It aims to be the foundational layer for the future of statistical computing in Python. It serves as a strong complement to the existing scientific Python stack while implementing and improving upon the kinds of data manipulation tools found in other statistical programming languages such as R. In addition to detailing its design and features of pandas, we will discuss future avenues of work and growth opportunities for statistics and data analysis applications in the Python language.},
   file = {/Users/martinkuric/Zotero/storage/IH5C5UZ3/§-mckinneyPandasFoundationalPython2011.pdf}
+}
+
+@misc{reback2020pandas,
+  title = {Pandas-Dev/Pandas: {{Pandas}}},
+  author = {pandas development {team}, The},
+  year = {2020},
+  month = feb,
+  doi = {10.5281/zenodo.3509134},
+  howpublished = {Zenodo}
 }
 
 @article{vallatPingouinStatisticsPython2018,
@@ -53,7 +69,6 @@
   pages = {1026},
   issn = {2475-9066},
   doi = {10.21105/joss.01026},
-  url = {https://joss.theoj.org/papers/10.21105/joss.01026},
   urldate = {2023-05-29},
   abstract = {Vallat, (2018). Pingouin: statistics in Python. Journal of Open Source Software, 3(31), 1026, https://doi.org/10.21105/joss.01026},
   langid = {english},
@@ -72,7 +87,6 @@
   pages = {3021},
   issn = {2475-9066},
   doi = {10.21105/joss.03021},
-  url = {https://joss.theoj.org/papers/10.21105/joss.03021},
   urldate = {2023-03-26},
   abstract = {Waskom, M. L., (2021). seaborn: statistical data visualization. Journal of Open Source Software, 6(60), 3021, https://doi.org/10.21105/joss.03021},
   langid = {english},
@@ -89,7 +103,6 @@
   pages = {1--23},
   issn = {1548-7660},
   doi = {10.18637/jss.v059.i10},
-  url = {https://doi.org/10.18637/jss.v059.i10},
   urldate = {2023-11-15},
[…]
