+++ paper.bib
@@ -30,6 +30,15 @@
   file = {/Users/martinkuric/Zotero/storage/W4FJZDNY/§-hunterMatplotlib2DGraphics2007.pdf;/Users/martinkuric/Zotero/storage/GW3HZZHR/4160265.html}
 }
 
+@inproceedings{mckinneyDataStructuresStatistical2010,
+  title = {Data {{Structures}} for {{Statistical Computing}} in {{Python}}},
+  author = {McKinney, Wes},
+  year = {2010},
+  month = jan,
+  pages = {56--61},
+  doi = {10.25080/Majora-92bf1922-00a}
+}
+
 @article{mckinneyPandasFoundationalPython2011,
   title = {Pandas: A {{Foundational Python Library}} for {{Data Analysis}} and {{Statistics}}},
   shorttitle = {Pandas},
@@ -39,6 +48,16 @@
   journal = {Python High Performance Science Computer},
   abstract = {---In this paper we will discuss pandas, a Python library of rich data structures and tools for working with structured data sets common to statistics, finance, social sciences, and many other fields. The library provides integrated, intuitive routines for performing common data manipulations and analysis on such data sets. It aims to be the foundational layer for the future of statistical computing in Python. It serves as a strong complement to the existing scientific Python stack while implementing and improving upon the kinds of data manipulation tools found in other statistical programming languages such as R. In addition to detailing its design and features of pandas, we will discuss future avenues of work and growth opportunities for statistics and data analysis applications in the Python language.},
   file = {/Users/martinkuric/Zotero/storage/IH5C5UZ3/§-mckinneyPandasFoundationalPython2011.pdf}
+}
+
+@misc{reback2020pandas,
+  title = {Pandas-Dev/Pandas: {{Pandas}}},
+  author = {The Pandas Development Team},
+  year = {2020},
+  month = feb,
+  doi = {10.5281/zenodo.3509134},
+  url = {https://doi.org/10.5281/zenodo.3509134},
+  howpublished = {Zenodo}
 }
 
 @article{vallatPingouinStatisticsPython2018,
