+++ paper.bib
@@ -5,6 +5,7 @@
   year = {2022},
   month = oct,
   doi = {10.5281/ZENODO.7213391},
+  url = {https://zenodo.org/record/7213391},
   urldate = {2023-11-16},
   abstract = {Add scipy's Brunner-Munzel test Fix applying statannotations for non-string group labels (Issue \#65) Get Zenodo DOI},
   copyright = {Open Access},
@@ -23,6 +24,7 @@
   pages = {90--95},
   issn = {1558-366X},
   doi = {10.1109/MCSE.2007.55},
+  url = {https://ieeexplore.ieee.org/document/4160265},
   urldate = {2023-11-15},
   abstract = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting,and publication-quality image generation across user interfaces and operating systems},
   file = {/Users/martinkuric/Zotero/storage/W4FJZDNY/§-hunterMatplotlib2DGraphics2007.pdf;/Users/martinkuric/Zotero/storage/GW3HZZHR/4160265.html}
@@ -70,6 +72,7 @@
   pages = {1026},
   issn = {2475-9066},
   doi = {10.21105/joss.01026},
+  url = {https://joss.theoj.org/papers/10.21105/joss.01026},
   urldate = {2023-05-29},
   abstract = {Vallat, (2018). Pingouin: statistics in Python. Journal of Open Source Software, 3(31), 1026, https://doi.org/10.21105/joss.01026},
   langid = {english},
@@ -88,6 +91,7 @@
   pages = {3021},
   issn = {2475-9066},
   doi = {10.21105/joss.03021},
+  url = {https://joss.theoj.org/papers/10.21105/joss.03021},
   urldate = {2023-03-26},
   abstract = {Waskom, M. L., (2021). seaborn: statistical data visualization. Journal of Open Source Software, 6(60), 3021, https://doi.org/10.21105/joss.03021},
   langid = {english},
@@ -104,6 +108,7 @@
   pages = {1--23},
   issn = {1548-7660},
   doi = {10.18637/jss.v059.i10},
+  url = {https://doi.org/10.18637/jss.v059.i10},
   urldate = {2023-11-15},
   abstract = {A huge amount of effort is spent cleaning data to get it ready for analysis, but there has been little research on how to make data cleaning as easy and effective as possible. This paper tackles a small, but important, component of data cleaning: data tidying. Tidy datasets are easy to manipulate, model and visualize, and have a specific structure: each variable is a column, each observation is a row, and each type of observational unit is a table. This framework makes it easy to tidy messy datasets because only a small set of tools are needed to deal with a wide range of un-tidy datasets. This structure also makes it easier to develop tidy tools for data analysis, tools that both input and output tidy datasets. The advantages of a consistent data structure and matching tools are demonstrated with a case study free from mundane data manipulation chores.},
   copyright = {Copyright (c) 2013 Hadley  Wickham},
