+++ paper.md
@@ -227,16 +227,9 @@ to override `plotastic` settings. Figures are exported using
 `plotastic` also focuses on annotating statistical information within
 plots, seamlessly incorporating p-values from pairwise comparisons using
 `statannotations` [@charlierTrevismdStatannotationsV02022]. This
-integration simplifies the interface and expands options for pair
+integration simplifies the interface and enables options for pair
 selection in multidimensional plots, enhancing both user experience and
 interpretability.
-
-`plotastic` also focuses on annotating statistical information inside
- the plot. It integrates `statannotations` for annotating p-values from
- pairwise comparisons inside the plot
- [@charlierTrevismdStatannotationsV02022], and simplifies the interface
- and offering further options for pair selection in multidimensional
- plots.
 
 For statistics, `plotastic` integrates with the `pingouin` library to
 support classical assumption and hypothesis testing, covering
